Jettison _verify_redis_connection
Broken in the worst possible way---it works, by accident. Mostly. And very
inefficiently.

Before every operation, it calls _verify_redis_connection, which pings the
Redis server. If successful, it sets $success to 1 and returns. But NOT from
_verify_redis_connection, just from the try block. (Even if it did return from
_verify_redis_connection, it would be returning false!)

So, it falls through to building a new Redis object. Every. Damned. Time.

If you happened to pass in a pre constructed Redis object and didn't provide a
server parameter, it either gets replaced with one constructed on localhost
with default arguments, or if you don't have a Redis server on localhost, it
fails to build and replace it. But since $success is set to true, and you
still have your original Redis object set, _verify_redis_connection returns
true and the operation continues.

Net effect, if you passed in a server parameter:
- ping the server
- create a new connection
- perform the operation

Net effect, if you passed in a pre-constructed Redis object:
- ping the server
- fail to create Redis object
- bitch about it (filling up log files)
- perform the operation